+++ src/RFBCM_CFD/mesh/include/meshData.hpp
@@ -20,34 +20,36 @@ class MeshData
     MeshData(std::shared_ptr<controlData> inControlData);
     ~MeshData(){};
 
-    // std::map<int, std::string> nodeToGoupMap_;
+    // std::map<size_t, std::string> nodeToGoupMap_;
     // std::map<std::string, boundaryCondition> groupToBoundaryConditionMap_;
-    // std::map<int, boundaryCondition> nodeToBoundaryConditionMap_;
+    // std::map<size_t, boundaryCondition> nodeToBoundaryConditionMap_;
 
     std::vector<vec3d<double>>& nodes();
-    int numOfNodes() const;
-    std::shared_ptr<BoundaryCondition> nodeBC(const int nodeID) const;
+    size_t numOfNodes() const;
+    std::shared_ptr<BoundaryCondition> nodeBC(const size_t nodeID) const;
 
-    nodesCloud neighborNodesCloud(const int nodeID, const int neighborNum);
+    nodesCloud neighborNodesCloud(const size_t nodeID,
+                                  const size_t neighborNum);
 
-    // std::shared_ptr<BoundaryCondition> nodesBC(const int nodeID) const;
+    // std::shared_ptr<BoundaryCondition> nodesBC(const size_t nodeID) const;
 
    private:
     std::shared_ptr<controlData> controlData_;
     std::vector<vec3d<double>> nodes_;
     std::vector<vec3d<double>> normals_;
 
-    std::map<std::string, std::vector<int>> groupToNodesMap_;
+    std::map<std::string, std::vector<size_t>> groupToNodesMap_;
     std::map<std::string, std::shared_ptr<BoundaryCondition>> groupToBCMap_;
     std::vector<std::shared_ptr<BoundaryCondition>> nodesToBC_;
     std::vector<std::string> nodesToGroup_;
 
-    int numOfNodes_;
+    size_t numOfNodes_;
     KDTreeEigenAdaptor<std::vector<vec3d<double>>, double, 3> kdTree_;
 
     void buildBoundaryConditions();
-    void compactGroupToNodesMap(const std::map<std::string, std::vector<int>>&
-                                    groupToNodesMapBeforeCompact);
+    void compactGroupToNodesMap(
+        const std::map<std::string, std::vector<size_t>>&
+            groupToNodesMapBeforeCompact);
 };
 
 #endif
