+++ src/RFBCM_CFD/mesh/include/rectangle.hpp
@@ -10,9 +10,9 @@
 class Rectangle
 {
    public:
-    int numInnNodes_, numBouNodes_, numAllNodes_;
+    size_t numInnNodes_, numBouNodes_, numAllNodes_;
 
-    Rectangle(int numNodeX, int numNodeY, double sizeX, double sizeY);
+    Rectangle(size_t numNodeX, size_t numNodeY, double sizeX, double sizeY);
 
     ~Rectangle(){};
 
@@ -20,15 +20,15 @@ class Rectangle
     const std::vector<std::vector<double> >& getNormals() const;
 
    private:
-    int numNodeX_, numNodeY_;
+    size_t numNodeX_, numNodeY_;
     double sizeX_, sizeY_;
     double centerX_, centerY_;
     std::vector<std::vector<double> > nodes_;
     std::vector<std::vector<double> > normals_;
 
-    std::vector<double> computeAllNodes(int i) const;
-    std::vector<double> computeInnerNodes(int i) const;
-    std::vector<double> computeBoundaryNodes(int i) const;
-    std::vector<double> computeNormals(int i) const;
+    std::vector<double> computeAllNodes(size_t i) const;
+    std::vector<double> computeInnerNodes(size_t i) const;
+    std::vector<double> computeBoundaryNodes(size_t i) const;
+    std::vector<double> computeNormals(size_t i) const;
 };
 #endif
